# public/swagger.yaml
openapi: 3.0.3
info:
  title: e-EnergieTEC AMI - NIGELEC Developer Portal
  description: |
    API de gestion de l'infrastructure de comptage intelligent pour la NIGELEC.
    Cette API permet d'interagir avec les compteurs STS, les concentrateurs DCU, le système de facturation et la télérelève DLMS/COSEM.
    
    ### Authentification
    Toutes les requêtes nécessitent un token JWT dans le header `Authorization`.
    Exemple : `Authorization: Bearer <your_jwt_token>`
  version: 2.5.0
  contact:
    name: Direction Informatique NIGELEC
    email: support.it@nigelec.ne
servers:
  - url: http://localhost:3000/api
    description: Serveur Local / Production NIGELEC (Port 3000)
  - url: https://api.nigelec.ne/v2
    description: Serveur de Production National HES

paths:
  /auth/login:
    post:
      tags:
        - Authentification
      summary: Authentification Opérateur / Agent & Obtention JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  example: admin123
      responses:
        '200':
          description: Jeton JWT généré avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    type: object

  /meters:
    get:
      tags:
        - Compteurs AMI
      summary: Liste complète des compteurs AMI installés
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [online, warning, offline]
      responses:
        '200':
          description: Liste des compteurs retournée avec succès
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Meter'

    post:
      tags:
        - Compteurs AMI
      summary: Enregistrer un nouveau compteur sur le réseau
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Meter'
      responses:
        '201':
          description: Compteur créé avec succès

  /meters/{id}:
    put:
      tags:
        - Compteurs AMI
      summary: Mettre à jour la fiche d'un compteur
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Meter'
      responses:
        '200':
          description: Mis à jour avec succès

    delete:
      tags:
        - Compteurs AMI
      summary: Supprimer un compteur du parc
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Compteur supprimé

  /meters/{id}/credit:
    post:
      tags:
        - STS Prepaid & Monétique
      summary: Générer un jeton de recharge STS (20 Digits)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
              properties:
                amount:
                  type: number
                  example: 5000
                tariffId:
                  type: string
                  example: TARIFF-DOM-01
      responses:
        '200':
          description: Jeton STS généré avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: "1234 5678 9012 3456 7890"
                  kwh:
                    type: number
                    example: 63.05

  /meters/{id}/sync-clock:
    post:
      tags:
        - Compteurs AMI
      summary: Synchroniser l'horloge interne DLMS du compteur
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Horloge synchronisée

  /dcus:
    get:
      tags:
        - Concentrateurs DCU
      summary: Liste des concentrateurs de données de collecte (PLC/RF/4G)
      responses:
        '200':
          description: Succès

    post:
      tags:
        - Concentrateurs DCU
      summary: Déployer un nouvel équipement DCU
      responses:
        '201':
          description: Concentrateur créé

  /dcus/{id}/ping:
    post:
      tags:
        - Concentrateurs DCU
      summary: Tester la télémesure (Ping HES) d'un DCU
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ping réussi

  /customers:
    get:
      tags:
        - Gestion Clients
      summary: Liste des abonnés NIGELEC rattachés aux compteurs
      responses:
        '200':
          description: Succès

  /tariffs:
    get:
      tags:
        - Tarification & ARSE
      summary: Grille tarifaire officielle d'électricité au Niger
      responses:
        '200':
          description: Grille tarifaire retournée

  /mdms/stats:
    get:
      tags:
        - Supervision & MDMS
      summary: Statistiques nationales de comptage et bilan énergétique
      responses:
        '200':
          description: Statistiques consolidées

  /audits:
    get:
      tags:
        - Sécurité & Audit
      summary: Journal d'audit immuable des opérations système
      responses:
        '200':
          description: Logs d'audit signés SHA256

components:
  schemas:
    Meter:
      type: object
      properties:
        id:
          type: string
          example: "0128260224778"
        serialNumber:
          type: string
          example: "0128260224778"
        location:
          type: string
          example: "Niamey / Koubia"
        type:
          type: string
          enum: [domestic, social, commercial, industrial, haute_tension, eclairage_public]
          example: "domestic"
        status:
          type: string
          enum: [online, warning, offline]
          example: "online"
        credit:
          type: number
          example: 42.5
        phaseType:
          type: string
          enum: [monophase, triphase]
          example: "monophase"
        power:
          type: number
          example: 1.5
        voltage:
          type: number
          example: 230

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - bearerAuth: []